EC/requirements.txt
aiofiles==24.1.0
asyncio==3.4.3
paho-mqtt==2.1.0
pycryptodome==3.20.0
websockets==13.0.1
